\name{statistical_rethinking-package}
\alias{statistical_rethinking-package}
\alias{statistical_rethinking}
\docType{package}
\title{
\packageTitle{statistical_rethinking}
}
\description{
\packageDescription{statistical_rethinking}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{statistical_rethinking}
\packageIndices{statistical_rethinking}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{statistical_rethinking}

Maintainer: \packageMaintainer{statistical_rethinking}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
